fix: Force region sizes to always be 64 bit

+++ lib/include/pl/helpers/types.hpp
@@ -19,7 +19,7 @@ namespace pl {
 
     struct Region {
         u64 address;
-        size_t size;
+        u64 size;
     };
 
     enum class ControlFlowStatement {
